+++ session_test.go
@@ -18,124 +18,12 @@ package main
 import (
 	"fmt"
 	"net/http"
-	"reflect"
 	"testing"
 	"time"
 
-	"github.com/coreos/go-oidc/jose"
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 )
-
-func getFakeAccessToken(t *testing.T) jose.JWT {
-	testToken, err := jose.NewJWT(
-		jose.JOSEHeader{
-			"alg": "RS256",
-		},
-		jose.Claims{
-			"jti": "4ee75b8e-3ee6-4382-92d4-3390b4b4937b",
-			//"exp": "1450372969",
-			"nbf":            0,
-			"iat":            "1450372669",
-			"iss":            "https://keycloak.example.com/auth/realms/commons",
-			"aud":            "test",
-			"sub":            "1e11e539-8256-4b3b-bda8-cc0d56cddb48",
-			"typ":            "Bearer",
-			"azp":            "clientid",
-			"session_state":  "98f4c3d2-1b8c-4932-b8c4-92ec0ea7e195",
-			"client_session": "f0105893-369a-46bc-9661-ad8c747b1a69",
-			"resource_access": map[string]interface{}{
-				"openvpn": map[string]interface{}{
-					"roles": []string{
-						"dev-vpn",
-					},
-				},
-			},
-			"email":              "gambol99@gmail.com",
-			"name":               "Rohith Jayawardene",
-			"family_name":        "Jayawardene",
-			"preferred_username": "rjayawardene",
-			"given_name":         "Rohith",
-		},
-	)
-	if err != nil {
-		t.Fatalf("unable to generate a token: %s", err)
-	}
-
-	return testToken
-}
-
-func getFakeRealmAccessToken(t *testing.T) jose.JWT {
-	testToken, err := jose.NewJWT(
-		jose.JOSEHeader{
-			"alg": "RS256",
-		},
-		jose.Claims{
-			"jti": "4ee75b8e-3ee6-4382-92d4-3390b4b4937b",
-			//"exp": "1450372969",
-			"nbf":            0,
-			"iat":            "1450372669",
-			"iss":            "https://keycloak.example.com/auth/realms/commons",
-			"aud":            "test",
-			"sub":            "1e11e539-8256-4b3b-bda8-cc0d56cddb48",
-			"typ":            "Bearer",
-			"azp":            "clientid",
-			"session_state":  "98f4c3d2-1b8c-4932-b8c4-92ec0ea7e195",
-			"client_session": "f0105893-369a-46bc-9661-ad8c747b1a69",
-			"realm_access": map[string]interface{}{
-				"roles": []string{
-					"dsp-dev-vpn",
-					"vpn-user",
-					"dsp-prod-vpn",
-				},
-			},
-			"resource_access": map[string]interface{}{
-				"openvpn": map[string]interface{}{
-					"roles": []string{
-						"dev-vpn",
-					},
-				},
-			},
-			"email":              "gambol99@gmail.com",
-			"name":               "Rohith Jayawardene",
-			"family_name":        "Jayawardene",
-			"preferred_username": "rjayawardene",
-			"given_name":         "Rohith",
-		},
-	)
-	if err != nil {
-		t.Fatalf("unable to generate a token: %s", err)
-	}
-
-	return testToken
-}
-
-func TestGetUserContext(t *testing.T) {
-	context, err := extractIdentity(getFakeAccessToken(t))
-	assert.NoError(t, err)
-	assert.NotNil(t, context)
-	assert.Equal(t, "1e11e539-8256-4b3b-bda8-cc0d56cddb48", context.id)
-	assert.Equal(t, "gambol99@gmail.com", context.email)
-	assert.Equal(t, "rjayawardene", context.preferredName)
-	roles := []string{"openvpn:dev-vpn"}
-	if !reflect.DeepEqual(context.roles, roles) {
-		t.Errorf("the claims are not the same, %v <-> %v", context.roles, roles)
-	}
-}
-
-func TestGetUserRealmRoleContext(t *testing.T) {
-	context, err := extractIdentity(getFakeRealmAccessToken(t))
-	assert.NoError(t, err)
-	assert.NotNil(t, context)
-	assert.Equal(t, "1e11e539-8256-4b3b-bda8-cc0d56cddb48", context.id)
-	assert.Equal(t, "gambol99@gmail.com", context.email)
-	assert.Equal(t, "rjayawardene", context.preferredName)
-	roles := []string{"dsp-dev-vpn", "vpn-user", "dsp-prod-vpn", "openvpn:dev-vpn"}
-	if !reflect.DeepEqual(context.roles, roles) {
-		t.Errorf("the claims are not the same, %v <-> %v", context.roles, roles)
-	}
-
-}
 
 func TestGetSessionToken(t *testing.T) {
 	token := getFakeAccessToken(t)
@@ -162,7 +50,7 @@ func TestGetSessionToken(t *testing.T) {
 				},
 			},
 		},
-		// @TODO need to ather checks
+		// @TODO need to other checks
 	}
 
 	for i, c := range testCases {
@@ -181,36 +69,33 @@ func TestGetSessionToken(t *testing.T) {
 }
 
 func TestEncodeState(t *testing.T) {
-	state := &refreshSession{
+	state := &RefreshToken{
 		token:    "this is a fake session",
 		expireOn: time.Now(),
 	}
-
-	session, err := encryptStateSession(state, "")
+	session, err := encryptRefreshToken(state, "1gjrlcjQ8RyKANngp9607txr5fF5fhf1")
 	assert.NotEmpty(t, session)
 	assert.NoError(t, err)
 }
 
 func TestDecodeState(t *testing.T) {
-	fakeKey := "HYLNt2JSzD7Lpz0djTRudmlOpbwx1oHBtr:"
-	fakeToken := "this is a fake session"
-	fakeExpiresOn := time.Now()
+	fakeKey := "HYLNt2JSzD7Lpz0djTRudmlOpbwx1oHB"
 
-	state := &refreshSession{
-		token:    fakeToken,
-		expireOn: fakeExpiresOn,
+	state := &RefreshToken{
+		token:    "this is a fake session",
+		expireOn: time.Now(),
 	}
 
-	session, err := encryptStateSession(state, fakeKey)
-	assert.NotEmpty(t, session)
-	if err != nil {
+	encrypted, err := encryptRefreshToken(state, fakeKey)
+	if !assert.NoError(t, err) {
 		t.Errorf("the encryptStateSession() should not have handed an error")
 		t.FailNow()
 	}
+	assert.NotEmpty(t, encrypted)
 
-	decoded, err := decryptRefreshSession(session, fakeKey)
+	decoded, err := decryptRefreshToken(encrypted, fakeKey)
 	assert.NotNil(t, decoded, "the session should not have been nil")
 	if assert.NoError(t, err, "the decodeState() should not have thrown an error") {
-		assert.Equal(t, fakeToken, decoded.token, "the token should been the same")
+		assert.Equal(t, fakeToken, decoded, "the token should been the same")
 	}
 }